refactor: move slots/message into utils directory
