+++ crypto/utils/slot.py
