+++ tests/test_slot.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from crypto.conf import use_network
-from crypto.slot import get_epoch, get_time
+from crypto.utils.slot import get_epoch, get_time
 
 
 def test_get_epoch():
